Make more direct assertion calls

+++ jgrapht-core/src/test/java/org/jgrapht/alg/color/LargestDegreeFirstColoringTest.java
@@ -72,11 +72,11 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(2, colors.get(2).intValue());
-        assertEquals(1, colors.get(3).intValue());
-        assertEquals(2, colors.get(4).intValue());
-        assertEquals(2, colors.get(5).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(2, colors.get(2));
+        assertEquals(1, colors.get(3));
+        assertEquals(2, colors.get(4));
+        assertEquals(2, colors.get(5));
     }
 
     @Test
@@ -98,12 +98,12 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(1, colors.get(1).intValue());
-        assertEquals(2, colors.get(2).intValue());
-        assertEquals(0, colors.get(3).intValue());
-        assertEquals(2, colors.get(4).intValue());
-        assertEquals(2, colors.get(5).intValue());
-        assertEquals(1, colors.get(6).intValue());
+        assertEquals(1, colors.get(1));
+        assertEquals(2, colors.get(2));
+        assertEquals(0, colors.get(3));
+        assertEquals(2, colors.get(4));
+        assertEquals(2, colors.get(5));
+        assertEquals(1, colors.get(6));
     }
 
     @Test
@@ -121,12 +121,12 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(0, colors.get(2).intValue());
-        assertEquals(1, colors.get(3).intValue());
-        assertEquals(2, colors.get(4).intValue());
-        assertEquals(0, colors.get(5).intValue());
-        assertEquals(1, colors.get(6).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(0, colors.get(2));
+        assertEquals(1, colors.get(3));
+        assertEquals(2, colors.get(4));
+        assertEquals(0, colors.get(5));
+        assertEquals(1, colors.get(6));
     }
 
     @Test
@@ -143,12 +143,12 @@ public class LargestDegreeFirstColoringTest
         Coloring<Integer> coloring = new LargestDegreeFirstColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(0, colors.get(2).intValue());
-        assertEquals(1, colors.get(3).intValue());
-        assertEquals(1, colors.get(4).intValue());
-        assertEquals(0, colors.get(5).intValue());
-        assertEquals(2, colors.get(6).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(0, colors.get(2));
+        assertEquals(1, colors.get(3));
+        assertEquals(1, colors.get(4));
+        assertEquals(0, colors.get(5));
+        assertEquals(2, colors.get(6));
     }
 
 }
