+++ jgrapht-core/src/test/java/org/jgrapht/alg/color/SmallestDegreeLastColoringTest.java
@@ -56,11 +56,11 @@ public class SmallestDegreeLastColoringTest
         Coloring<Integer> coloring = new SmallestDegreeLastColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(2, colors.get(1).intValue());
-        assertEquals(0, colors.get(2).intValue());
-        assertEquals(1, colors.get(3).intValue());
-        assertEquals(0, colors.get(4).intValue());
-        assertEquals(0, colors.get(5).intValue());
+        assertEquals(2, colors.get(1));
+        assertEquals(0, colors.get(2));
+        assertEquals(1, colors.get(3));
+        assertEquals(0, colors.get(4));
+        assertEquals(0, colors.get(5));
     }
 
     @Test
@@ -83,12 +83,12 @@ public class SmallestDegreeLastColoringTest
         Coloring<Integer> coloring = new SmallestDegreeLastColoring<>(g).getColoring();
         assertEquals(3, coloring.getNumberColors());
         Map<Integer, Integer> colors = coloring.getColors();
-        assertEquals(0, colors.get(1).intValue());
-        assertEquals(1, colors.get(2).intValue());
-        assertEquals(0, colors.get(3).intValue());
-        assertEquals(2, colors.get(4).intValue());
-        assertEquals(1, colors.get(5).intValue());
-        assertEquals(0, colors.get(6).intValue());
+        assertEquals(0, colors.get(1));
+        assertEquals(1, colors.get(2));
+        assertEquals(0, colors.get(3));
+        assertEquals(2, colors.get(4));
+        assertEquals(1, colors.get(5));
+        assertEquals(0, colors.get(6));
     }
 
 }
